Increase Node.js memory limit
The amount of data in WOF is reaching the point where some workers are
hitting the default Node.js memory limit and crashing. Increasing the
Node.js limit doesn't fix the problem, but gives us time to come up with
a new architecture that does.

Connected to https://github.com/pelias/wof-admin-lookup/issues/117